AssignmentFunction
minor changes
