+++ Skolpatrullen/WebApp/Controllers/AssignmentController.cs
@@ -2,15 +2,26 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Database.Models;
 using Microsoft.AspNetCore.Mvc;
 
 namespace WebApp.Controllers
 {
-    public class AssignmentController : Controller
+    public class AssignmentController : AppController
     {
-        public IActionResult Index()
+        [HttpPost]
+        [Route("[controller]/AddCourseAssignment")]
+        public async Task<IActionResult> AddCourseAssignment(Assignment assignment, int courseId)
         {
-            return View();
+            string message = await GetUser();
+            if (!ModelState.IsValid)
+            {
+                return View();
+            }
+            assignment.CourseId = courseId;
+            var response = await APIAddAssignment(assignment);
+
+            return RedirectToAction("GetCourseById", "Course", new { Id = assignment.CourseId });
         }
     }
 }
