+++ Skolpatrullen/WebApp/Views/Course/_AddCourseAssignment.cshtml
@@ -31,7 +31,7 @@
             </div>
             <div class="modal-footer">
                 <button type="button" class="btn btn-secondary" data-dismiss="modal">Avbryt</button>
-                <input type="submit" asp-action="AddCourseAssignment" asp-route-courseId="@Model.CourseId" value="Spara" class="btn btn-primary" />
+                <input type="submit" asp-controller="Assignment" asp-action="AddCourseAssignment" asp-route-courseId="@Model.CourseId" value="Spara" class="btn btn-primary" />
             </div>
         </form>
     </div>
