+++ Skolpatrullen/WebApp/WebApp.csproj
@@ -19,6 +19,10 @@
     <ProjectReference Include="..\Lib\Lib.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Views\Assignment\" />
+  </ItemGroup>
+
 
 
 </Project>
